refactor(tools): extract erc20 module, deduplicate fetch_protocol_systems
- Move ERC-20 utilities (slot detection, allowance reading) into
  tools/fynd-swap-cli/src/erc20.rs; move their unit tests there too
- Extract fetch_protocol_systems into fynd-rpc/src/protocols.rs,
  eliminating the duplicate copy in src/main.rs
- Rename TransferType variants to TransferFrom/TransferFromPermit2
  to mirror UserTransferType from fynd-rpc-types
