+++ src/main.rs
@@ -36,12 +36,8 @@ use clap::Parser;
 use fynd_rpc::{
     builder::{parse_chain, FyndBuilder},
     config::{defaults, BlacklistConfig, WorkerPoolsConfig},
+    protocols::fetch_protocol_systems,
 };
-use tycho_simulation::{
-    tycho_client::rpc::{HttpRPCClient, HttpRPCClientOptions, RPCClient},
-    tycho_common::models::Chain,
-};
-
 mod cli;
 use cli::{Cli, Commands};
 use metrics_exporter_prometheus::{PrometheusBuilder, PrometheusHandle};
@@ -53,7 +49,7 @@ use tokio::{
     select,
     signal::unix::{signal, SignalKind},
 };
-use tracing::{debug, error, info, warn};
+use tracing::{error, info, warn};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter};
 
 fn main() -> Result<(), anyhow::Error> {
@@ -355,42 +351,3 @@ async fn run_solver(args: cli::ServeArgs) -> Result<(), SolverError> {
     Ok(())
 }
 
-/// Fetches all available protocol systems from Tycho RPC.
-async fn fetch_protocol_systems(
-    tycho_url: &str,
-    auth_key: Option<&str>,
-    use_tls: bool,
-    chain: Chain,
-) -> Result<Vec<String>, anyhow::Error> {
-    use tycho_simulation::tycho_common::dto::{PaginationParams, ProtocolSystemsRequestBody};
-
-    info!("Fetching available protocol systems from Tycho RPC...");
-
-    let rpc_url =
-        if use_tls { format!("https://{tycho_url}") } else { format!("http://{tycho_url}") };
-    let rpc_options = HttpRPCClientOptions::new().with_auth_key(auth_key.map(|s| s.to_string()));
-    let rpc_client = HttpRPCClient::new(&rpc_url, rpc_options)?;
-
-    const PAGE_SIZE: i64 = 100;
-    let mut all_protocols = Vec::new();
-    let mut page = 0;
-
-    loop {
-        let request = ProtocolSystemsRequestBody {
-            chain: chain.into(),
-            pagination: PaginationParams { page, page_size: PAGE_SIZE },
-        };
-        let response = rpc_client
-            .get_protocol_systems(&request)
-            .await?;
-        let count = response.protocol_systems.len();
-        all_protocols.extend(response.protocol_systems);
-        if (count as i64) < PAGE_SIZE {
-            break;
-        }
-        page += 1;
-    }
-
-    debug!("Fetched {} protocol system(s) from Tycho RPC", all_protocols.len());
-    Ok(all_protocols)
-}
